requirements.txt
RPi.GPIO
telepot
